+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - Added a podcast general settings page
 - Added an option to disallow publishing RSS for a podcast
-- Each podcast has two settings pages (general settings, episode settings)
+- Now each podcast has two settings pages (general settings, episode settings)
 - Fixed a bug where changing lightswitch inputs on the sidebar of the episode edit page, creates unwanted drafts
+- Fixed permission issues where non-admin users were not able to access some plugin actions
 
 ## 0.1.0 - 2023-03-02